refactor: Update scheduling block priority type to f64 and rename SkyMapBlock to LightweightBlock for improved performance and clarity

+++ rust_backend/src/parsing/json_parser.rs
@@ -143,7 +143,7 @@ fn parse_scheduling_block(
     let priority = value
         .get("priority")
         .and_then(|v| v.as_f64())
-        .context("Missing or invalid 'priority'")? as f32;
+        .context("Missing or invalid 'priority'")? ;
 
     // Extract target coordinates
     let target = parse_target(value)?;
